# internal/handler/static/openapi.yaml
openapi: "3.0.3"
info:
  title: Aletheia API
  version: "0.2.0"
  summary: Certificação de mídia ancorada em blockchain.
  description: |
    A Aletheia API certifica imagens calculando um SHA-256 do
    conteúdo, ancorando esse hash em um contrato EVM e armazenando
    metadados em PostgreSQL. Também extrai uma assinatura visual
    (pHash + descritores ORB) que permite verificação por
    similaridade quando o arquivo foi re-encodado, rotacionado,
    espelhado ou recortado levemente.

    ## Fluxos principais

    - `POST /certificates` registra uma imagem nova. O SHA-256 é a
      chave única; reenvio do mesmo arquivo devolve `409 Conflict`.
    - `GET /certificates/verify?hash=...` consulta por hash exato.
    - `POST /certificates/verify` aceita upload e tenta primeiro o
      match exato por SHA-256, depois o caminho de similaridade
      visual via pHash + ORB.

    ## Convenções

    - Todo corpo é JSON (`application/json`), exceto uploads que usam
      `multipart/form-data` com o campo `file` (apenas imagens).
    - Hashes SHA-256 são strings hex de 64 caracteres em minúsculas.
    - Datas são RFC 3339 em UTC (`2006-01-02T15:04:05Z`).
    - Erros seguem o schema `Error` com um único campo `error`.

    ## Autenticação

    Três níveis, deliberadamente distintos:

    - **Público** — sem credencial. Alcança `/health`, `/docs` e a
      verificação. A verificação é a metade gratuita do produto e
      nunca é bloqueada.
...
      resposta que a criou.
...

...
  contact:
...
servers:
...
paths:
...
  /certificates:
    post:
      tags: [Certificates]
      summary: Certifica um upload sem atestação (legado)
      description: |
        Rota de migração para integrações anteriores ao SDK.

        Fica **desligada** salvo `ALLOW_UNATTESTED_CERTIFY=true`, e
        responde `410 Gone` quando desligada. Um upload que ninguém
        endossou é exatamente o que a captura atestada substitui:
        deixá-la ligada permitiria cunhar certificados que parecem
        oficiais mas não provam nada sobre origem.

        O certificado resultante traz `attested: false`.
      operationId: certifyContent
      security: [{ TenantKey: [] }]
      parameters:
        - $ref: "#/components/parameters/RegistrantHeader"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadRequest"
            encoding:
              file:
                contentType: >-
                  image/jpeg, image/png, image/gif, image/webp,
                  image/bmp, image/tiff
      responses:
        "201":
          description: Conteúdo certificado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
              examples:
                imagem:
                  $ref: "#/components/examples/CertificateImage"
        "400":
          $ref: "#/components/responses/BadFile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Conteúdo já certificado (mesmo SHA-256).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                duplicado:
                  value: { error: "certify: content already certified" }
        "410":
          description: Uploads não atestados estão desligados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                desligado:
                  value: { error: "unattested uploads are disabled; certify through POST /captures with an enrolled device" }
        "413":
          description: Arquivo excede o limite de upload (100 MB).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          $ref: "#/components/responses/UnsupportedMedia"
        "422":
          description: Falha ao processar (erro de extração ou persistência).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /certificates/verify:
    get:
      tags: [Certificates]
      summary: Verifica certificação por hash
      description: |
        Consulta direta por `SHA-256`. Útil quando o cliente já tem o
        hash e só quer confirmar se foi certificado e por quem.

        Retorna `200 OK` com o certificado quando encontrado e
        `404 Not Found` com `certified=false` quando não há registro.
      operationId: verifyByHash
      parameters:
        - $ref: "#/components/parameters/HashQuery"
      responses:
        "200":
          description: Conteúdo certificado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
              examples:
                certificado:
                  $ref: "#/components/examples/VerifyHit"
        "400":
          description: Parâmetro `hash` ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                vazio:
                  value: { error: "query parameter 'hash' is required" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Nenhum certificado para o hash informado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
              examples:
                nao_certificado:
                  $ref: "#/components/examples/VerifyMiss"

    post:
      tags: [Certificates]
      summary: Verifica certificação por upload de imagem
      description: |
        Aceita um upload de imagem e tenta dois caminhos em sequência:

        1. Match exato por `SHA-256`.
        2. Extrai pHash em várias rotações/flips, faz prefilter LSH
           em `phash_bands`, calcula Hamming-256 contra os candidatos
           sobreviventes (`maxDistance` configurado em domínio) e,
           nos melhores `top-K`, roda re-check com ORB + resíduo de
           cor contra a grade de cores armazenada no certificado.

        O custo do segundo caminho cresce com o tamanho da base e da
        imagem enviada, então prefira o `GET ?hash=` quando o
        cliente já tem o hash.
      operationId: verifyByFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadRequest"
            encoding:
              file:
                contentType: >-
                  image/jpeg, image/png, image/gif, image/webp,
                  image/bmp, image/tiff
      responses:
        "200":
          description: Conteúdo certificado (match exato ou por similaridade).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
              examples:
                exato:
                  $ref: "#/components/examples/VerifyHit"
                similar:
                  $ref: "#/components/examples/VerifyHitSimilar"
        "400":
          $ref: "#/components/responses/BadFile"
        "404":
          description: Nenhum certificado correspondente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
              examples:
                nao_certificado:
                  $ref: "#/components/examples/VerifyMiss"
        "413":
          description: Arquivo excede o limite de upload (100 MB).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          $ref: "#/components/responses/UnsupportedMedia"
        "500":
          description: Erro interno (leitura, extração ou banco).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                interno:
                  value: { error: "verify: postgres find by hash: connection refused" }
  /certificates/{hash}:
    delete:
      tags: [Certificates]
      summary: Remove um certificado
      description: |
        Apaga o certificado identificado pelo `SHA-256`. Todos os
        dados (assinatura ORB, grade de cores, bandas de pHash) vivem
        no banco, então a remoção é uma única operação atômica.

        Responde `204 No Content` em caso de sucesso (sem corpo) e
        `404 Not Found` quando não há certificado para o hash.
      operationId: deleteCertificate
      security: [{ AdminToken: [] }]
      parameters:
        - $ref: "#/components/parameters/HashPath"
      responses:
        "204":
          description: Certificado removido. Sem corpo.
        "404":
          description: Nenhum certificado para o hash informado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                nao_encontrado:
                  value: { error: "delete: certificate not found" }
        "500":
          description: Erro interno (falha na remoção no banco).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                interno:
                  value: { error: "delete: postgres delete: timeout" }
security: []
components:
  securitySchemes:
    TenantKey:
      type: http
      scheme: bearer
      description: |
        Chave de API de uma organização (`alk_…`). Toda falha —
        malformada, desconhecida, revogada, organização suspensa —
        devolve o mesmo `401`, para que a resposta não sirva de sonda
        sobre quais credenciais existem.
    AdminToken:
      type: http
      scheme: bearer
      description: |
        `ADMIN_API_TOKEN`. Em branco **rejeita tudo** em vez de
        desligar o guarda.
  parameters:
    RegistrantHeader:
      in: header
      name: X-Registrant
      required: false
      schema:
        type: string
        maxLength: 256
      description: |
        Identificador do registrante (endereço de carteira EVM,
        identificador organizacional ou similar). É gravado como está
        no certificado e devolvido em consultas. Hoje **não** é
        validado pela API.
      examples:
        wallet:
          value: "0x742d35Cc6634C0532925a3b844Bc9e7595f2bD18"
        org:
          value: "newsroom:globo"
    HashQuery:
      in: query
      name: hash
      required: true
      schema:
        $ref: "#/components/schemas/Sha256Hex"
      description: SHA-256 do conteúdo, em hex de 64 caracteres.
      example: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
    HashPath:
      in: path
      name: hash
      required: true
      schema:
        $ref: "#/components/schemas/Sha256Hex"
      description: SHA-256 do conteúdo a remover, em hex de 64 caracteres.
      example: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
  responses:
    Unauthorized:
      description: |
        Credencial ausente, inválida, revogada ou de organização
        suspensa. Todas colapsam na mesma resposta de propósito.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            tenant:
              value: { error: "a valid API key is required" }
            admin:
              value: { error: "admin credentials required" }
    BadFile:
      description: Campo `file` ausente, vazio ou inválido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            ausente:
              value: { error: "missing or invalid file field" }
    UnsupportedMedia:
      description: >-
        Tipo MIME não aceito. Aceitos: `image/jpeg`, `image/png`,
        `image/gif`, `image/webp`, `image/bmp`, `image/tiff`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            tipo_invalido:
              value: { error: "only image files are accepted" }
  schemas:
    Sha256Hex:
      type: string
      pattern: "^[0-9a-f]{64}$"
      minLength: 64
      maxLength: 64
      description: Hash SHA-256 em hex minúsculo, 64 caracteres.
      example: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
...
    EvmTxHash:
      type: string
      pattern: "^0x[0-9a-fA-F]{64}$"
      description: Hash de transação EVM (32 bytes hex com prefixo `0x`).
      example: "0x9fce0c2b9b0d2d8f3a1e5e7d4c5b3a8f7e6d4c3b2a1908f7e6d5c4b3a2918e7c"
    PendingEvmTxHash:
      type: string
      pattern: "^$|^0x[0-9a-fA-F]{64}$"
      description: |
        Hash de transação EVM, ou string vazia enquanto o certificado
        aguarda o próximo lote de ancoragem. Certificar não toca a
        blockchain, então todo certificado recém-criado começa aqui.
      example: "0x9fce0c2b9b0d2d8f3a1e5e7d4c5b3a8f7e6d4c3b2a1908f7e6d5c4b3a2918e7c"
    UploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: |
            Imagem a certificar/verificar. Limite de 100 MB. Tipos
            aceitos: JPEG, PNG, GIF, WebP, BMP, TIFF.
    Certificate:
      type: object
      required:
        - id
        - content_hash
        - registrant
        - tx_hash
        - block_number
        - created_at
        - attested
      properties:
        id:
          type: string
          format: uuid
          description: Identificador interno do certificado.
          example: "550e8400-e29b-41d4-a716-446655440000"
        content_hash:
          $ref: "#/components/schemas/Sha256Hex"
        registrant:
          type: string
          description: |
            Organização de registro. Em capturas atestadas é o id da
            organização autenticada; no caminho legado anônimo é o
            valor recebido em `X-Registrant`, que nunca foi evidência
            de nada.
          example: "9f1c8d2e-4b6a-4c3d-8e7f-1a2b3c4d5e6f"
        attested:
          type: boolean
          description: |
            Se o conteúdo veio de uma captura por dispositivo inscrito
            (`true`) ou de um upload comum (`false`). É o campo mais
            importante para quem verifica: decide o quanto o
            certificado vale.
          example: true
        device_id:
          type: string
          format: uuid
          description: Dispositivo que capturou. Ausente em uploads não atestados.
          example: "7b2e4f10-9c3d-4a5b-8e6f-0a1b2c3d4e5f"
        captured_at:
          type: string
          format: date-time
          description: |
            Instante de captura reportado pelo dispositivo, coberto
            pela assinatura. Ausente em uploads não atestados.
          example: "2026-08-12T15:04:05.123456789Z"
        tx_hash:
          allOf:
            - $ref: "#/components/schemas/PendingEvmTxHash"
          description: |
            Transação que ancorou o lote deste certificado. Vazia
            enquanto o certificado aguarda o próximo lote — use o
            objeto `anchor`, que só aparece quando a ancoragem
            confirmou, para saber se há prova de inclusão.
        block_number:
          type: integer
          format: int64
          minimum: 0
          description: |
            Bloco que incluiu a transação, obtido do receipt — sempre
            um bloco real. Zero enquanto não ancorado.
          example: 63481022
        anchor:
          allOf:
            - $ref: "#/components/schemas/Anchor"
          description: |
            Compromisso on-chain. Ausente enquanto o certificado
            aguarda o próximo lote.
        created_at:
          type: string
          format: date-time
          description: Instante de criação em UTC (RFC 3339).
          example: "2026-08-12T15:04:06Z"
    Anchor:
      type: object
      description: |
        Prova de inclusão do certificado na raiz Merkle ancorada.
        Permite conferir contra a raiz on-chain sem confiar nesta API:
        recompute a folha como
        `keccak256(keccak256(contentHash ‖ featureCommitment))` e
        chame `AnchorRegistry.verify(root, leaf, proof)`.
      required: [tx_hash, block_number, leaf_index, merkle_proof]
      properties:
        tx_hash:
          $ref: "#/components/schemas/EvmTxHash"
        block_number:
          type: integer
          format: int64
          example: 63481022
        leaf_index:
          type: integer
          minimum: 0
          description: Posição do certificado no lote.
          example: 17
        merkle_proof:
          type: array
          description: Hashes irmãos, da folha para cima, em hex 0x.
          items:
            type: string
            pattern: "^0x[a-f0-9]{64}$"
...
    VerifyResponse:
      type: object
      required: [certified]
      properties:
        certified:
          type: boolean
          description: |
            `true` se há certificado correspondente, `false` caso
            contrário. Quando `false`, `certificate` virá ausente ou
            nulo e o status HTTP será `404`.
          example: true
        certificate:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Certificate"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Mensagem legível descrevendo a falha.
          example: "missing or invalid file field"
  examples:
    CertificateImage:
      summary: Certificado de imagem com âncora on-chain
      value:
        id: "550e8400-e29b-41d4-a716-446655440000"
        content_hash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
        registrant: "0x742d35Cc6634C0532925a3b844Bc9e7595f2bD18"
        tx_hash: "0x9fce0c2b9b0d2d8f3a1e5e7d4c5b3a8f7e6d4c3b2a1908f7e6d5c4b3a2918e7"
        block_number: 0
        created_at: "2026-02-25T12:00:00Z"
    VerifyHit:
      summary: Match exato por SHA-256
      value:
        certified: true
        certificate:
          id: "550e8400-e29b-41d4-a716-446655440000"
          content_hash: "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
          registrant: "0x742d35Cc6634C0532925a3b844Bc9e7595f2bD18"
          tx_hash: "0x9fce0c2b9b0d2d8f3a1e5e7d4c5b3a8f7e6d4c3b2a1908f7e6d5c4b3a2918e7"
          block_number: 0
          created_at: "2026-02-25T12:00:00Z"
    VerifyHitSimilar:
      summary: Match por similaridade visual (pHash + ORB)
      value:
        certified: true
        certificate:
          id: "a1b2c3d4-e5f6-7890-abcd-ef0123456789"
          content_hash: "1f0e2d3c4b5a69788796a5b4c3d2e1f00f1e2d3c4b5a69788796a5b4c3d2e1f0"
          registrant: "newsroom:globo"
          tx_hash: "0x1111111111111111111111111111111111111111111111111111111111111111"
          block_number: 0
          created_at: "2026-02-20T08:13:42Z"
    VerifyMiss:
      summary: Conteúdo não certificado
      value:
        certified: false
        certificate: null